requirements.txt
pillow 
python-dotenv 
langchain-google-genai
streamlit